// frontend/person-blog-app/packages/blog/src/shared/api/generated/models/blogModel.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Gateway.API
 * OpenAPI spec version: 1.0
 */

export interface BlogModel {
  id?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  createdAt?: string;
  /** @nullable */
  photoUrl?: string | null;
  userId?: string;
  subscribersCount?: number;
}